requirements.txt
pymongo==4.6.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
beautifulsoup4==4.12.2
requests==2.31.0
snscrape==0.7.0.20230622
praw==7.7.1
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1
wordcloud==1.9.3
pyvi==0.1.1
underthesea==6.7.0
scikit-learn==1.3.2
streamlit==1.29.0
python-dotenv==1.0.0